+++ misc/highlight.js
@@ -6,28 +6,7 @@
 // all integers are LE
 //
 
-var hljs = require('highlight.js'),
-    escape = require('escape-html');
-
-global.hljs = hljs;
-
-// Custom HLJS langs
-require('./hljs/rodacode');
-
-function highlight(string, language, langId) {
-    var head = '<pre class="hljs exec-target" data-lang="' + langId + '"><code>',
-        tail = '</code></pre>';
-
-    if (language && hljs.getLanguage(language)) {
-        // try {
-            return head + hljs.highlight(language, string, true).value + tail;
-        // } catch (__) {
-        //     return '<div style="font-size: 3rem; color: red; font-weight: 300">Error Rendering</div>';
-        // }
-    }
-
-    return head + escape(string) + tail;
-}
+const highlight = require('./hljs-renderer');
 
 process.stdin.resume();
