+++ misc/hljs/rodacode.js
@@ -1,1 +1,1 @@
-hljs.registerLanguage("rodacode",function(e){var t={keyword:"and boolean break catch continue do done else floating for function if in integer is list map new not number or record reference reflect return string try typeof unless until value while xor",literal:"E ENV FALSE PI STDERR STDIN STDOUT TRUE",built_in:"abs acos addHead addTail asin assignGlobal assignGlobalType assignType atan atan2 bufferedExec bytesToString cbrt cd ceil chars chr concat connect cos cosh count createGlobal createGlobalType createType currentTime enum error errprint exec exp exportGlobal exportGlobalType false fileExists fileLength filePermissions filter floor getLine getenv grep head identity ifStdIn ifStdOut import importNamespace indexOf inputMode interleave isDirectory isFile jlineVar json keys kwColor ln loadResourceLines localImport localImportNamespace log log10 lowerCase ls makeSeq match max mimeType min name open ord orderedUniq parseFloating parseInteger peek print product prompt pull push pwd randomBoolean randomFloating randomInteger randomize readLines reduce reduceSteps replace require reverse round safeLocalImport saveResource search seq server shift signum sin sinh slide sort split splitMany sqrt stream stringToBytes strsize sum tail tan tanh thread true tryPeek tryPull undefine uniq unorderedCount unorderedUniq unpull upperCase varColor writeStrings"};return{aliases:["rodalang", "roda", "roeda"],k:t,i:"</",c:[e.CBCM,{cN:"string",v:[e.QSM]},{cN:"number",v:[e.CNM]},{b:/:=/},{cN:'meta',b:'@[A-Za-z]+'},{cN:"function",bK:"function",e:/\s*\{|\n/,eE:!0,c:[e.TM,{cN:"params",b:/\(/,e:/\)/,k:t,i:/["]/}]}]}});
+module.exports=function(e){var t={keyword:"and boolean break catch continue do done else floating for function if in integer is list map new not number or record reference reflect return string try typeof unless until value while xor",literal:"E ENV FALSE PI STDERR STDIN STDOUT TRUE",built_in:"abs acos addHead addTail asin assignGlobal assignGlobalType assignType atan atan2 bufferedExec bytesToString cbrt cd ceil chars chr concat connect cos cosh count createGlobal createGlobalType createType currentTime enum error errprint exec exp exportGlobal exportGlobalType false fileExists fileLength filePermissions filter floor getLine getenv grep head identity ifStdIn ifStdOut import importNamespace indexOf inputMode interleave isDirectory isFile jlineVar json keys kwColor ln loadResourceLines localImport localImportNamespace log log10 lowerCase ls makeSeq match max mimeType min name open ord orderedUniq parseFloating parseInteger peek print product prompt pull push pwd randomBoolean randomFloating randomInteger randomize readLines reduce reduceSteps replace require reverse round safeLocalImport saveResource search seq server shift signum sin sinh slide sort split splitMany sqrt stream stringToBytes strsize sum tail tan tanh thread true tryPeek tryPull undefine uniq unorderedCount unorderedUniq unpull upperCase varColor writeStrings"};return{aliases:["rodalang", "roda", "roeda"],k:t,i:"</",c:[e.CBCM,{cN:"string",v:[e.QSM]},{cN:"number",v:[e.CNM]},{b:/:=/},{cN:'meta',b:'@[A-Za-z]+'},{cN:"function",bK:"function",e:/\s*\{|\n/,eE:!0,c:[e.TM,{cN:"params",b:/\(/,e:/\)/,k:t,i:/["]/}]}]}};
